ai-engine/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pypdf==4.2.0
python-docx==1.1.2
numpy==1.26.4
sentence-transformers
scikit-learn
python-dotenv
google-genai